codes/data/requirements.txt
hydra-core==0.11.3
Faker==3.0.0
pySSHA==0.6
